perf: add LLMDerivService as the base of LLMDerivs
