+++ Easydict/Swift/Service/LLMDerivs/PolishingService.swift
@@ -9,35 +9,12 @@
 import Foundation
 
 @objc(EZPolishingService)
-class PolishingService: BuiltInAIService {
-    // MARK: Public
-
+class PolishingService: LLMDerivService {
     public override func name() -> String {
         NSLocalizedString("polishing_service", comment: "")
     }
 
     public override func serviceType() -> ServiceType {
         .polishing
-    }
-
-    public override func configurationListItems() -> Any {
-        StreamConfigurationView(
-            service: self,
-            showNameSection: false,
-            showAPIKeySection: false,
-            showEndpointSection: false,
-            showSupportedModelsSection: false,
-            showUsedModelSection: false,
-            showTranslationToggle: false,
-            showSentenceToggle: false,
-            showDictionaryToggle: false,
-            showUsageStatusPicker: true
-        )
-    }
-
-    // MARK: Internal
-
-    override var defaultModels: [String] {
-        ["llama3-70b-8192"]
     }
 }
